event-consumer/requirements.txt
pika
couchdb
python-dotenv